+++ data/src/main/resources/db/migration/V2__add_osm_node_id.sql
@@ -1,0 +1,1 @@
+ALTER TABLE pos ADD COLUMN osm_node_id bigint UNIQUE;